# specs/001-ai-consultation-platform/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Car Service AI Assistant API
  version: 0.5.0
  description: >
    Полный операционный контракт, сгенерированный из runtime Express
    (`docs/api-route-inventory.json`). Схемы тел запросов — минимальные;
    источник истины по наличию path+method — инвентарь + `npm run openapi:check`.
servers:
  - url: http://localhost:3000/api
...
paths:
...
  "/vehicles":
    get:
      tags: [Vehicles]
      summary: Получить vehicles
      operationId: get_vehicles
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      tags: [Vehicles]
      summary: Создать/выполнить vehicles
      operationId: post_vehicles
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/vehicles/{vehicleId}":
    delete:
      tags: [Vehicles]
      summary: "Удалить {vehicleId}"
      operationId: delete_vehicles_vehicleId
      parameters:
        - in: path
          name: vehicleId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    get:
      tags: [Vehicles]
      summary: "Получить {vehicleId}"
      operationId: get_vehicles_vehicleId
      parameters:
        - in: path
          name: vehicleId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    patch:
      tags: [Vehicles]
      summary: "Обновить {vehicleId}"
      operationId: patch_vehicles_vehicleId
      parameters:
        - in: path
          name: vehicleId
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  "/vehicles/{vehicleId}/photo":
    delete:
      tags: [Vehicles]
      summary: Удалить photo
      operationId: delete_vehicles_vehicleId_photo
      parameters:
        - in: path
          name: vehicleId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    get:
      tags: [Vehicles]
      summary: Получить photo
      operationId: get_vehicles_vehicleId_photo
      parameters:
        - in: path
          name: vehicleId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      tags: [Vehicles]
      summary: Создать/выполнить photo
      operationId: post_vehicles_vehicleId_photo
      parameters:
        - in: path
          name: vehicleId
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        error: { type: string }
        code: { type: string }
...